fixed dtype mismatch in `_paste_mask`

+++ official/vision/beta/projects/panoptic_maskrcnn/modeling/layers/panoptic_segmentation_generator.py
@@ -92,7 +92,8 @@ class PanopticSegmentationGenerator(tf.keras.layers.Layer):
     super(PanopticSegmentationGenerator, self).__init__(**kwargs)
 
   def _paste_mask(self, box, mask):
-    pasted_mask = tf.ones(self._output_size + [1]) * self._void_class_label
+    pasted_mask = tf.ones(
+        self._output_size + [1], dtype=mask.dtype) * self._void_class_label
 
     ymin = box[0]
     xmin = box[1]
